don't re-install!

+++ containers/postgis-root-scripts/osm.bash
@@ -1,20 +1,26 @@
 #! /bin/bash
 
-apt-get update
-apt-get install -qqy --no-install-recommends \
-  osm2pgsql \
-  osmctools \
-  osmium-tool \
-  osmosis
-apt-get clean \
-apt-file update \
-update-command-not-found
+if [ ! -e /usr/bin/osmosis ]
+then
+  apt-get update
+  apt-get install -qqy --no-install-recommends \
+    osm2pgsql \
+    osmctools \
+    osmium-tool \
+    osmosis
+  apt-get clean \
+  apt-file update \
+  update-command-not-found
+fi
 
-cd /usr/local/src
-rm -fr osm2pgrouting-*
-curl -Ls https://github.com/pgRouting/osm2pgrouting/archive/v2.3.3.tar.gz | tar xzf -
-cd osm2pgrouting-*
-cmake -H. -Bbuild
-cd build
-make
-make install
+if [ ! -e /usr/local/bin/osm2pgrouting ]
+then
+  cd /usr/local/src
+  rm -fr osm2pgrouting-*
+  curl -Ls https://github.com/pgRouting/osm2pgrouting/archive/v2.3.3.tar.gz | tar xzf -
+  cd osm2pgrouting-*
+  cmake -H. -Bbuild
+  cd build
+  make
+  make install
+fi
